chore: folders testing and router
Folders creation and relocation files
Added jest and testing-library packages
Modified files about app routing
Created .babelrc file
Added exeptions in .gitignore file
