+++ .gitignore
@@ -26,6 +26,9 @@ chrome-user-data
 dist
 .eslintcache
 
+assets/
+data/
+
 packages/react-devtools-core/dist
 packages/react-devtools-extensions/chrome/build
 packages/react-devtools-extensions/chrome/*.crx
