+++ src/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import reactDOM from 'react-dom';
-import App from './App';
+import App from './routes/App';
 
 const root = document.getElementById('root');
 reactDOM.render(<App />, root);
