--- src/pages/NotFound.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
-
-const NotFound = () => (
-    <div>
-        <h3>Not Found Page</h3>
-        <Link to="/">Go home</Link>
-    </div>
-);
-
-export default NotFound;
