+++ src/pages/Theme/themeExample.scss
@@ -1,0 +1,4 @@
+.content {
+    display: grid;
+    gap: 20px;
+}
