--- src/style/themeExample.scss
@@ -1,4 +1,0 @@
-.content {
-    display: grid;
-    gap: 20px;
-}
